Fix #746 project file removed if save error.
This could occur in two places. First, Util::warnIfNotWritable() removes
the file if the write test succeeds. Obviously, if the eventual save
does not take place, then nothing will exist. This is fine for something
like a job that has a deferred action and might be removed. However, for
something like saving to an existing file whose save operation might
fail, this is bad. In the case of saving XML, the generated XML might be
malformed, or Shotuct could crash during the save operation.

The second case is in QFile::rename() according to its docs. This could
occur because the temporary file is locked, for example. In that case,
we should have a safety net: Instead of removing an existing destination
file prior to rename, rename the existing destination file with a name
that has "backup" in it (in the same folder).

+++ src/util.cpp
@@ -78,7 +78,7 @@ void Util::showInFolder(const QString& path)
     QDesktopServices::openUrl(QUrl::fromLocalFile(info.isDir()? path : info.path()));
 }
 
-bool Util::warnIfNotWritable(const QString& filePath, QWidget* parent, const QString& caption)
+bool Util::warnIfNotWritable(const QString& filePath, QWidget* parent, const QString& caption, bool remove)
 {
     // Returns true if not writable.
     if (!filePath.isEmpty()) {
@@ -93,7 +93,7 @@ bool Util::warnIfNotWritable(const QString& filePath, QWidget* parent, const QSt
                                     "Try again with a different folder.")
                                  .arg(fi.fileName()));
             return true;
-        } else {
+        } else if (remove) {
             file.remove();
         }
     }
